allow specifying the location of the long description help text
this text can go at the top (as it always has) or at the bottom after the full command and argument printout. Since the help text often gets extremely lengthy any long description is often out of sight above the top of the screen. specifying the location allows this text to be easier to see